castform: Download Environment Canada Weather Station Data

Downloads hourly weather station data (as csv files) from Environment Canada.

Main functions

get_metadata: Downloads and loads the latest station inventory metadata (RUN THIS FIRST)

get_single_station_file: The primary function to download one hourly data file.

get_multiple_station_files: Downloads multiple hourly data files.

province_station_files: Downloads data files by province.

year_range_station_files: Downloads data files by year range.

get_all_files: Downloads all data files with hourly data available.

station_lookup: Search for available station data by province and year range.

Author

Maintainer: Shawn Yip shawniceyip@trentu.ca