castform: Download Environment Canada Weather Station Data

Downloads hourly weather station data (as csv files) from Environment Canada. This data is downloaded from: https://climate.weather.gc.ca/historical_data/search_historic_data_e.html

Metadata

get_metadata: Downloads and loads the latest station inventory metadata (RUN THIS FIRST)

station_lookup: Search for available station data by province and year range.

Download Wrappers

get_single_station_file: The primary function to download one hourly data file.

get_multiple_station_files: Downloads multiple hourly data files.

get_station_files: Downloads data files by station

province_station_files: Downloads data files by province.

year_range_station_files: Downloads data files by year range.

get_all_files: Downloads all data files with hourly data available.

Database Building

build_station_database: Creates a database using a folder of specified weather station data

validate_database: Validates created database

Exploratory Data Analysis

station_map: Creates a map visualizing stations of interest

data_missingness_table: Creates a table summarizing actual, expected, and missing data counts

data_ranges: Creates a table summarizing variable data ranges

plot_yearly_means: Creates plots to summarize yearly variable means

pull_missing_strings: Creates a plot and table summarizing missing data gaps

pull_repeated_strings: Creates a plot and table summarizing repeated data strings

Heatwave Detector

heatwave_detector: Identifies extreme heat events (heatwaves) based on set temperature thresholds.

See also

Author

Maintainer: Shawn Yip shawniceyip@trentu.ca